zephyr: Add and run "outputexports" buildsystem target.
This target is used to export CC, CFLAGS, etc. definitions from Zephyr to
external build systems. This is the same approach as used by JerryScript,
MicroPython, open62541, etc. ports to Zephyr. This target used to be part
of Zephyr build system, but it regressed at time of migrating it to CMake.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>