libiotc/iotc.c: Don't try to print uint32_t as int
The code assume that uint32_t has an "int" underlying type, but that
doesn't have to be true. E.g., with Zephyr toolchains, uint32_t has
long base type, so such usage produces warning-as-error and breaks the
build.

This works it around ny casting to unsigned int, though the proper fix
would be to use PRIu32 stdint formatting code.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>